+++ MLS/MLS/SharedViews/CustomTextField.swift
@@ -9,30 +9,6 @@ import UIKit
 
 import SnapKit
 
-enum TextState: String {
-    case `default`
-    case complete
-    case emailCheck = "이메일을 다시 확인해주세요."
-    case emailBlank = "이메일을 입력해주세요."
-    case pwCheck = "비밀번호를 다시 확인해주세요."
-    case pwBlank = "비밀번호를 입력해주세요."
-    case pwOutOfBounds
-    case pwNotENG
-    case pwNotInt
-    case pwNotSymbol
-    case pwNotCorrect = "비밀번호가 일치하지 않아요"
-    case emailExist = "이미 가입된 이메일이에요."
-    case nickNameExist = "중복된 닉네임이에요."
-    case nickNameNotCorrect = "닉네임을 2~8자로 지어주세요"
-    case lvNotInt = "숫자만 입력해주세요"
-    case lvOutOfBounds = "1~200 사잇값만 넣어주세요"
-}
-
-enum TextFieldType {
-    case normal
-    case password
-}
-
 class CustomTextField: UIStackView {
     // MARK: - Properties
     
@@ -200,10 +176,10 @@ extension CustomTextField {
     func setPasswordFooter(checkPassword: [Bool], state: TextState) {
         footerLabel.isHidden = false
 
-        let greenAttributes: [NSAttributedString.Key: Any] = [.foregroundColor: UIColor.semanticColor.text.success_bold]
-        let redAttributes: [NSAttributedString.Key: Any] = [.foregroundColor: UIColor.semanticColor.text.distructive_bold]
+        let greenAttributes: [NSAttributedString.Key: Any] = [.foregroundColor: UIColor.semanticColor.text.success_bold ?? UIColor.white]
+        let redAttributes: [NSAttributedString.Key: Any] = [.foregroundColor: UIColor.semanticColor.text.distructive_bold ?? UIColor.white]
         
-        for i in 0...3 {
+        for i in 0 ... 3 {
             switch i {
             case 0:
                 guard let range = attributedString.string.range(of: "8자리 이상") else { return }
@@ -245,5 +221,12 @@ extension CustomTextField {
         footerLabel.textColor = .semanticColor.text.secondary
         footerLabel.text = "숫자 1~200 사이의 값만 넣어주세요"
         footerLabel.isHidden = false
+    }
+}
+
+extension CustomTextField {
+    enum TextFieldType {
+        case normal
+        case password
     }
 }
